+++ cmd/internal/switcher/sonic/redis/db/db.go
@@ -1,13 +1,26 @@
 package db
 
-import "github.com/metal-stack/metal-core/cmd/internal/switcher/sonic/redis"
+type Config struct {
+	Databases map[string]database `json:"DATABASES"`
+	Instances map[string]instance `json:"INSTANCES"`
+}
+
+type database struct {
+	Id        int    `json:"id"`
+	Instance  string `json:"instance"`
+	Separator string `json:"separator"`
+}
+
+type instance struct {
+	Addr string `json:"unix_socket_path"`
+}
 
 type DB struct {
 	Config *ConfigDB
 	State  *StateDB
 }
 
-func New(cfg *redis.Config) *DB {
+func New(cfg *Config) *DB {
 	configDB := cfg.Databases["CONFIG_DB"]
 	stateDB := cfg.Databases["STATE_DB"]
